Add GET endpoint for content

+++ pages/api/content.ts
@@ -1,4 +1,5 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
+import { encrypt } from '@/lib/cryptography';
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 type Data = any;
@@ -15,9 +16,23 @@ export default async function handler(
     req: NextApiRequest,
     res: NextApiResponse<Data | Error>
 ) {
-    if (req.method == "DELETE") {
-        const col_id = req.query.coluuid;
-        const col_key = req.headers.key;
+    const col_id = req.query.coluuid;
+    const col_key = req.headers.key;
+    if (!col_id || !col_key) return res.status(400).json({ error: "Missing collection ID or key" });
+    if(encrypt(col_id as string) != col_key) return res.status(401).json({ error: "Invalid key" });
+    if (req.method == "GET") {
+        const url = `https://api.estuary.tech/collections/${col_id}?dir=${req.query.dir}`;
+        const result = await fetch(url, {
+            method: 'GET',
+            headers: {
+                "Authorization": "Bearer " + process.env.ESTUARY_API_KEY,
+            }
+        });
+        const data = await result.json();
+        if (result.status !== 200) return res.status(500).json(data);
+        res.status(200).json(data);
+    }
+    else if (req.method == "DELETE") {
         const url = `https://api.estuary.tech/collections/${col_id}/contents`;
         const result = await fetch(url, {
             method: 'DELETE',
